requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.46
Flask-Migrate==4.1.0
alembic==1.18.4
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
bcrypt==5.0.0
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
streamlit
pandas
Flask-Mail
